Add visualization for comparision values
