+++ frontend/src/components/Journey/comparisons/ComparisonMenPerProvince.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import '../../../assets/css/journey/journey.css';
+import CompVisualization from './compVisualization/CompVisualization';
 
 class ComparisonMenPerProvince extends React.Component {
 
@@ -58,7 +59,7 @@ class ComparisonMenPerProvince extends React.Component {
         <div className="journey_content">
           <p> <span className="bold red"> {(this.state.value) ? this.state.value : "Loading "}%</span> of <span className="red bold">men</span> have depression in <span className="red bold">{this.props.province}</span></p>
           <p>Year: {this.state.dataFromYear}</p>
-
+          <p>{this.state.value == "" ? "" : <CompVisualization percent={this.state.value} />}</p>
           <p>
             <button type="button" className="redButtonLink" onClick={() => this.props.prev()}>
               <i className="fa fa-angle-left bold"></i> Go back
